De-bouncing
Code to de-bounce the alarm relay.  This is mostly to combat the flashing that will occur if the alarm is tripped.  Note that it should set the thermostat back to a normal run state if the alarm is tripped.  Not ideal, but better than the thermostat constantly being changed while the relay flashes back and forth.