Two pin implementation
Use NO and NC contacts to determine alarm armed/disarmed status.

Try to find the lowest daily tstat setting for a few times with a delay in between before giving up.

Text me a message if the tstat setback fails so I can set it back using the app on my smartphone.